Move calculateColorLuminance, calculatePublishedDate and buildVTTFileLocally out of the store
