+++ src/renderer/views/ProfileEdit/ProfileEdit.js
@@ -5,6 +5,7 @@ import FtProfileEdit from '../../components/ft-profile-edit/ft-profile-edit.vue'
 import FtProfileChannelList from '../../components/ft-profile-channel-list/ft-profile-channel-list.vue'
 import FtProfileFilterChannelsList from '../../components/ft-profile-filter-channels-list/ft-profile-filter-channels-list.vue'
 import { MAIN_PROFILE_ID } from '../../../constants'
+import { calculateColorLuminance } from '../../helpers/utils'
 
 export default Vue.extend({
   name: 'ProfileEdit',
@@ -60,7 +61,7 @@ export default Vue.extend({
     if (profileType === 'newProfile') {
       this.isNew = true
       const bgColor = await this.getRandomColor()
-      const textColor = await this.calculateColorLuminance(bgColor)
+      const textColor = calculateColorLuminance(bgColor)
       this.profile = {
         name: '',
         bgColor: bgColor,
@@ -88,8 +89,7 @@ export default Vue.extend({
   methods: {
     ...mapActions([
       'showToast',
-      'getRandomColor',
-      'calculateColorLuminance'
+      'getRandomColor'
     ])
   }
 })
